+++ crates/bcr-ebill-wasm/src/data/mod.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use bcr_ebill_api::{
     data::{
         File, GeneralSearchFilterItemType, GeneralSearchResult, OptionalPostalAddress,
-        PostalAddress, UploadFilesResult,
+        PostalAddress, UploadFileResult,
     },
     util::file::{UploadFileHandler, detect_content_type_for_bytes},
 };
@@ -254,13 +254,13 @@ impl UploadFileHandler for UploadFile {
 
 #[derive(Tsify, Debug, Serialize, Clone)]
 #[tsify(into_wasm_abi)]
-pub struct UploadFilesResponse {
+pub struct UploadFileResponse {
     pub file_upload_id: String,
 }
 
-impl IntoWeb<UploadFilesResponse> for UploadFilesResult {
-    fn into_web(self) -> UploadFilesResponse {
-        UploadFilesResponse {
+impl IntoWeb<UploadFileResponse> for UploadFileResult {
+    fn into_web(self) -> UploadFileResponse {
+        UploadFileResponse {
             file_upload_id: self.file_upload_id,
         }
     }